add nerfstudio changes, change logics for caculating metrics for mulitple models, add render wrap

+++ src/subsample_data.py
@@ -1,11 +1,14 @@
 import argparse
-from pathlib import Path
-from enum import Enum
-import numpy as np
+import copy
+import json
 import operator
 import shutil
-from tqdm import tqdm
+from enum import Enum
+from pathlib import Path
 from shutil import copytree, ignore_patterns
+
+import numpy as np
+from tqdm import tqdm
 
 
 class Method(Enum):
@@ -56,26 +59,38 @@ def parse_args():
         default=[Method.DISPERSE],
     )
 
+    args.add_argument(
+        "--dry",
+        action="store_true",
+        help="Perform dry run",
+        default=False,
+    )
     return args.parse_known_args()
 
 
 def main():
     args, _ = parse_args()
+    is_dry_run = args.dry
     IMAGES_DIR_NAME = "images"
+    TRANSFORM_JSON_NAME = "transforms.json"
 
     for input_dir in args.input:
         input_dir = Path(input_dir)
         input_image_dir = Path(input_dir, IMAGES_DIR_NAME)
         input_dir_name = input_dir.name
 
-        sequence = sorted([f for f in input_image_dir.iterdir() if f.is_file()])
-        sequence_size = len(sequence)
+        image_paths = sorted([f for f in input_image_dir.iterdir() if f.is_file()])
+
+        sequence_size = len(image_paths)
+        zfill_size = len(str(sequence_size))
 
         subset_sizes = []
         subset_sizes.extend(
             [round(float(fraction) * sequence_size) for fraction in args.fractions]
         )
         subset_sizes.extend(args.ns)
+
+        transform_json = json.load(open(Path(input_dir, TRANSFORM_JSON_NAME), "r"))
 
         for method in args.methods:
             for subset_size in tqdm(subset_sizes):
@@ -92,23 +107,52 @@ def main():
                     indices_picked = np.round(fracs * (sequence_size - 1)).astype(int)
 
                 sequence_subset = operator.itemgetter(*(indices_picked.tolist()))(
-                    sequence
+                    image_paths
                 )
-                subset_dir_name = Path(args.output, f"{input_dir_name}_n_{subset_size}")
+
+                image_names_subset = sorted([path.stem for path in sequence_subset])
+
+                subset_dir_name = Path(
+                    args.output,
+                    f"{input_dir_name}_n_{str(subset_size).zfill(zfill_size)}",
+                )
                 subset_dir_name.mkdir(parents=True, exist_ok=True)
 
                 # Copy colmap and transforms
-                copytree(
-                    input_dir,
-                    subset_dir_name,
-                    ignore=ignore_patterns(IMAGES_DIR_NAME),
-                    dirs_exist_ok=True,
-                )
+                # copytree(
+                #     input_dir,
+                #     subset_dir_name,
+                #     ignore=ignore_patterns(IMAGES_DIR_NAME),
+                #     dirs_exist_ok=True,
+                # )
+
+                # Pick a subset of all frames
+                transform_json_subset = copy.deepcopy(transform_json)
+                transform_json_subset["frames"] = []
+                for frame in transform_json["frames"]:
+
+                    image_path = Path(frame["file_path"])
+                    image_name = image_path.stem
+                    if image_name in image_names_subset:
+                        transform_json_subset["frames"].append(frame)
+
+                if not transform_json_subset["frames"]:
+                    print("There are no frames in subset")
+                    exit(1)
+
+                new_transform_json_path = Path(subset_dir_name, TRANSFORM_JSON_NAME)
+
+                print("Creating transforms.json:", str(new_transform_json_path))
+
+                if not is_dry_run:
+                    with open(new_transform_json_path, "w") as f:
+                        json.dump(transform_json_subset, f)
 
                 for image in sequence_subset:
                     images_dir = Path(subset_dir_name, IMAGES_DIR_NAME)
-                    images_dir.mkdir(parents=True, exist_ok=True)
-                    shutil.copy(image, images_dir)
+                    if not is_dry_run:
+                        images_dir.mkdir(parents=True, exist_ok=True)
+                        shutil.copy(image, images_dir)
                     print("Saved to:", str(Path(images_dir, image.name)))
